toroplus: enable APN settings for CDMA
Traditionally, the APN settings menu is hidden on CDMA phones.
For legacy devices, this is not ideal due to many of them being
used on non-standard networks.  In previous versions of Android,
APN settings could be accessed by directly launching the activity,
but this no longer works in Marshmallow as the APNs are completely
hidden.

Enable the APN settings menu, which makes its contents visible to
the user.

Change-Id: Ia6d2bf4cc7c61308a770980ddf9b43dfecd7fc4e